Rename file document-uploader to document-to-base-64
